Remove old webhook to send plan changes and rely on events
